requirements.txt
aiogram==2.25.1
python-dotenv-1.0.0